+++ source/javascripts/services/workflows-selection-service.ts
@@ -1,6 +1,7 @@
 import { Step, Workflow } from "../models";
+import { WorkflowSelectionStore } from "./workflow-selection-store";
 
-type WfChainWrapper = {
+export type WfChainWrapper = {
 	workflow: Workflow;
 	isBeforeRunWorkflow?: boolean;
 	isAfterRunWorkflow?: boolean;
@@ -17,13 +18,6 @@ type WorkflowViewModel = {
 	selectedWorkflowChain: Array<WfChainWrapper>;
 };
 
-type WorkflowSelectionState = {
-	lastSelectedWorkflow?: Workflow;
-	lastEditedWorkflow?: Workflow;
-	lastEditedWorkflowIndex?: number;
-	lastSelectedStep?: Step;
-};
-
 type AngularLocationService = {
 	search: () => { workflow_id: string };
 };
@@ -36,68 +30,6 @@ const wfChainWrapper = (wrapper: WfChainWrapper): WfChainWrapper => ({
 	selectedWorkflowAfterRunWorkflowIndex: -1,
 	...wrapper
 });
-
-class WorkflowSelectionStore {
-	private shouldStoreChanges = false;
-
-	lastSelectedWorkflowID: string | null = null;
-	lastEditedWorkflowID: string | null = null;
-	lastEditedWorkflowIndex: number | null = null;
-	lastSelectedStepCVS: string | null = null;
-	lastSelectedStepIndex: number | null = null;
-
-	applyState = ({
-		lastEditedWorkflow,
-		lastSelectedStep,
-		lastSelectedWorkflow,
-		lastEditedWorkflowIndex
-	}: WorkflowSelectionState): void => {
-		if (!this.shouldStoreChanges) {
-			return;
-		}
-
-		if (lastSelectedWorkflow) {
-			this.lastSelectedWorkflowID = lastSelectedWorkflow.id;
-		}
-
-		if (lastEditedWorkflow && lastEditedWorkflowIndex) {
-			this.lastEditedWorkflowID = lastEditedWorkflow.id;
-			this.lastEditedWorkflowIndex = lastEditedWorkflowIndex;
-		}
-
-		if (lastSelectedStep) {
-			this.lastSelectedStepCVS = lastSelectedStep.cvs;
-		}
-
-		if (lastEditedWorkflow && lastSelectedStep) {
-			this.lastSelectedStepIndex = lastEditedWorkflow.steps.indexOf(lastSelectedStep);
-		}
-	};
-
-	checkLastSelectedWorkflow = (workflow: WfChainWrapper): boolean => {
-		return workflow.workflow?.id === this.lastEditedWorkflowID;
-	};
-
-	checkLastSelectedstep = (step: Step): boolean => {
-		return step.cvs === this.lastSelectedStepCVS;
-	};
-
-	reset = (): void => {
-		this.lastSelectedWorkflowID = null;
-		this.lastEditedWorkflowID = null;
-		this.lastEditedWorkflowIndex = null;
-		this.lastSelectedStepCVS = null;
-		this.lastSelectedStepIndex = null;
-	};
-
-	enable = (): void => {
-		this.shouldStoreChanges = true;
-	};
-
-	disable = (): void => {
-		this.shouldStoreChanges = false;
-	};
-}
 
 class WorkflowsSelectionService {
 	private store: WorkflowSelectionStore;
@@ -203,9 +135,5 @@ class WorkflowsSelectionService {
 	};
 }
 
-export const workflowSelectionStore = new WorkflowSelectionStore();
-
-export const serviceFactory = (
-	store: WorkflowSelectionStore,
-	location: AngularLocationService
-): WorkflowsSelectionService => new WorkflowsSelectionService(store, location);
+export default (store: WorkflowSelectionStore, location: AngularLocationService): WorkflowsSelectionService =>
+	new WorkflowsSelectionService(store, location);
